Rename repository class and the package name
